refactor: move model interfaces to preferences.model.ts and clean up imports in services
